refactor: move routers into api subpackages
