+++ app/domain/settings/api/__init__.py
@@ -1,0 +1,4 @@
+from .router import router
+
+__all__ = ['router']
+
