+++ app/domain/users/__init__.py
@@ -1,3 +1,3 @@
-from .router import router
+from .api.router import router
 
 __all__ = ["router"]
